Improve CLI capsule tests docstrings
Move the test modules to the right location. Improve docstrings
formatting and use @test as first docstring's line.

Also import CLITestCase which is the base test case class. And add the
new modules to the docs.

This is a follow up for the #1825
